new changes

+++ app/src/main/java/com/example/gonative/AuthViewModel.kt
@@ -104,9 +104,9 @@ class AuthViewModel : ViewModel() {
                             val rating = when (val value = it["rating"]) {
                                 is Double -> value.toFloat()
                                 is Int -> value.toFloat()
-                                is Long -> value.toFloat()
+                                is Long -> value.toLong()
                                 is String -> {
-                                    Log.d("Firebase", "Rating value: $value")  // Log the raw value
+                                    Log.d("Firebase", "Rating value: $value")
                                     value.toFloatOrNull() ?: 0.0f
                                 }
